+++ pleiades/__init__.py
@@ -1,8 +1,7 @@
 from pleiades.mesh import *
-from pleiades.fieldmath import *
 from pleiades.checkvalue import *
 from pleiades.transforms import *
-from pleiades.mixin import *
+from pleiades.fields import *
 from pleiades.current_sets import *
 from pleiades.device import *
 from pleiades.io import *
